delete done

+++ public/session/reservations.js
@@ -104,14 +104,25 @@ async function deleteReservation(){
     console.log(customer.children[1].innerHTML)
     let number = customer.children[1].innerHTML
     console.log(number)
-    let statusCode = await fetch("/api/customers/removeCustomer/"+number, {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
-    })
-    let elem = document.getElementById(number);
-    elem.parentNode.removeChild(elem);
-    let elem2 = document.getElementById("selectedCustomer")
-    elem2.innerHTML=""
+
+    var c = confirm("Vil du slette "+customer.children[0].innerHTML+"?")
+    if(c==true){
+        let statusCode = await fetch("/reservations/removeCustomer/"+number, {
+            method: "DELETE",
+            headers: { "Content-Type": "application/json" },
+        })
+        if(statusCode.status===200){
+        let elem = document.getElementById(number);
+        elem.parentNode.removeChild(elem);
+        let elem2 = document.getElementById("selectedCustomer")
+        elem2.innerHTML=""
+        } else {
+            console.log("delete error")
+        }
+        
+    } else {
+        // nothing
+    }
 
 }
